% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aion-package.R
\docType{package}
\name{aion-package}
\alias{aion}
\alias{aion-package}
\title{aion: Time Series Models with 'keras'}
\description{
Delivers useful functions to prepare input/ouput data for 'keras' time series models. Includes a 'Temporal Fusion Transformer' implementation and exposes its componets by sharing component layers in public package API. Introduces interface for 'keras-tcn' and 'keras_lmu' libraries and adds three loss function classes and one metric class.
}
\author{
Krzysztof Joachimiak
}
\keyword{internal}
