% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-3d-array.R
\name{as_3d_array}
\alias{as_3d_array}
\title{Create a 3-dimensional array out of a data.frame}
\usage{
as_3d_array(data, index, key)
}
\arguments{
\item{data}{A data.frame object}

\item{index}{A time-related column}

\item{key}{}
}
\description{
Create a 3-dimensional array out of a data.frame
}
\examples{
library(tsibbledata)
library(dplyr)

# We'll use the `global_economy` dataset
global_economy

selected_ge <-
 global_economy \%>\%
 select(Country, Year, Imports, Exports)



}
