% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dict_size}
\alias{dict_size}
\title{Get cardinalities of the categorical variables}
\usage{
dict_size(data, categorical)
}
\arguments{
\item{data}{A \code{data.frame} object}

\item{categorical}{List of the categorical variables}
}
\description{
Needed as an input to the \verb{[layer_multi_embedding()]}
}
\examples{
dict_size(m5::tiny_m5, c('event_name_1', 'event_type_1'))
}
