% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer-multi-dense.R
\name{layer_multi_dense}
\alias{layer_multi_dense}
\title{Multiple embeddings in one layer}
\usage{
layer_multi_dense(object, units, ...)
}
\arguments{
\item{object}{What to compose the new \code{Layer} instance with. Typically a
Sequential model or a Tensor (e.g., as returned by \code{layer_input()}).
The return value depends on \code{object}. If \code{object} is:
\itemize{
\item missing or \code{NULL}, the \code{Layer} instance is returned.
\item a \code{Sequential} model, the model with an additional layer is returned.
\item a Tensor, the output tensor from \code{layer_instance(object)} is returned.
}}

\item{units}{Positive integer, dimensionality of the output space.}
}
\description{
Multiple embeddings in one layer
}
