% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer-temporal-fusion-decoder.R
\name{layer_temporal_fusion_decoder}
\alias{layer_temporal_fusion_decoder}
\title{Temporal Fusion Decoder layer}
\usage{
layer_temporal_fusion_decoder(object)
}
\description{
Temporal Fusion Decoder layer
}
