% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer-vsn.R
\name{layer_vsn}
\alias{layer_vsn}
\title{Variable Selection Network block}
\usage{
layer_vsn(object, hidden_dim, state_size, dropout_rate = NULL, ...)
}
\arguments{
\item{dropout_rate}{Droput rate}
}
\description{
Variable Selection Network block
}
