% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss-quantile.R
\name{loss_quantile}
\alias{loss_quantile}
\title{Quantile loss function}
\usage{
loss_quantile(...)
}
\description{
A generalized version of the quantile loss. The model can predict multiple
quantiles at once. Loss value for a single sample-timestep-quantile is computed as:
}
\details{
\deqn{QL(y_t, y^{\hat}_t, q) = max(q(y - y^{\hat}))}
}
\examples{
y_pred <- array(runif(60), c(2, 10, 3))
y_true <- array(runif(20), c(2, 10, 1))

# As object
loss_quantile(quantiles = c(0.1, 0.5, 0.9), reduction = 'auto')(y_true, y_pred)

}
\references{
https://www.evergreeninnovations.co/blog-quantile-loss-function-for-machine-learning/
}
