% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-arrays.R
\name{make_arrays}
\alias{make_arrays}
\title{Prepare arrays (tensors), which can be used a}
\usage{
make_arrays(
  data,
  key,
  index,
  lookback,
  horizon,
  stride = 1,
  shuffle = TRUE,
  sample_frac = 1,
  target,
  numeric = NULL,
  categorical = NULL,
  static = NULL,
  past = NULL,
  future = NULL,
  by_key = TRUE
)
}
\arguments{
\item{data}{A data.frame object}

\item{key}{Columns, which creates unique time series IDs}

\item{index}{A time-related column}

\item{lookback}{The length of the context from the past}

\item{horizon}{The forecast length}

\item{stride}{Stride of the moving window}

\item{shuffle}{Shuffle samples. Set \code{FALSE} for the test dataset.}

\item{target}{Target variable(s)}

\item{numeric}{Numeric variables}

\item{categorical}{Categorical variables}

\item{static}{Static variables}
}
\description{
Prepare arrays (tensors), which can be used a
}
\examples{
library(m5)

make_arrays(
  data = tiny_m5,
  key = c('item_id', 'store_id'),
  index = 'date',
  lookback = 28,
  horizon = 28
)

}
