% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tft.R
\name{model_tft}
\alias{model_tft}
\title{Temporal Fusion Transformer model}
\usage{
model_tft(...)
}
\arguments{
\item{lookback}{Number of timesteps from the past}

\item{horizon}{Forecast length (number of timesteps)}

\item{past_numeric_size}{Number of numeric features from the past}

\item{past_categorical_size}{Number of categorical features from the past}

\item{future_numeric_size}{Number of numeric features from the future}
}
\description{
Temporal Fusion Transformer model
}
\examples{
library(keras)
library(aion)

tft <- model_tft(
   lookback                = 28,
   horizon                 = 14,
   past_numeric_size       = 5,
   past_categorical_size   = 2,
   future_numeric_size     = 4,
   future_categorical_size = 2,
   vocab_static_size       = c(5, 5),
   vocab_dynamic_size      = c(4, 4),
   optimizer               = 'adam',
   hidden_dim              = 12,
   state_size              = 7,
   n_heads                 = 10,
   dropout_rate            = 0.1,
   output_size             = 3
   #quantiles               = 0.5
)

X_static_cat <- array(sample(5, 32 * 2, replace=TRUE), c(32, 2)) - 1
X_static_num <- array(runif(32 * 1), c(32, 1))

X_past_num <- array(runif(32 * 28 * 2), c(32, 28, 2))
X_past_cat <- array(sample(4, 32 * 28 * 2, replace=TRUE), c(32, 28, 5))

X_fut_num <- array(runif(32 * 14 * 5), c(32, 28, 1))
X_fut_cat <- array(sample(4, 32 * 14 * 2, replace=TRUE), c(32, 28, 5))

tft(X_past_num, X_past_cat, X_fut_num, X_fut_cat, X_static_num, X_static_cat)

}
\references{
\enumerate{
\item \href{https://arxiv.org/abs/1912.09363}{Paper}
\item \href{https://github.com/google-research/google-research/blob/master/tft/libs/tft_model.py}{Original TFT implementation in TensorFlow}
\item \href{https://github.com/PlaytikaResearch/tft-torch/blob/main/tft_torch/tft.py}{A very clear implementation in PyTorch}
}
}
\seealso{
TFT components:
\code{\link[=layer_glu]{layer_glu()}}, \code{\link[=layer_grn]{layer_grn()}}, \code{\link[=layer_multi_embedding]{layer_multi_embedding()}}, \code{\link[=layer_multi_dense]{layer_multi_dense()}}
}
