% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neaten.R
\name{neaten}
\alias{neaten}
\title{Construct a data.frame out of the output forecast tensor}
\usage{
neaten(forecast, data, key, index, lookback, horizon, target, stride)
}
\arguments{
\item{data}{A data.frame containing test data}

\item{lookback}{The length of the context from the past}

\item{horizon}{The forecast length}

\item{target}{Target variable(s)}

\item{stride}{Stride of the moving window}

\item{x}{An \code{array} or \code{tensorflow.tensor} containing forecasts}
}
\description{
Construct a data.frame out of the output forecast tensor
}
