% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rand_array}
\alias{rand_array}
\title{Generate an array of random values with the given dimensions}
\usage{
rand_array(...)
}
\arguments{
\item{...}{Array dimensions.}
}
\value{
An array of random values between 0 and 1 with the defined dimensions.
}
\description{
Useful when testing developed model. We can check if the model accepts
input arrays of the expected size and if it returns output with desired shape.
}
\examples{
rand_array(2,3,4)
dim(rand_array(32,14,8))
}
