% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aion-package.R
\docType{package}
\name{aion-package}
\alias{aion}
\alias{aion-package}
\title{aion: Time Series Models with 'keras'}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\author{
Krzysztof Joachimiak
}
\keyword{internal}
