% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grn.R
\name{layer_grn}
\alias{layer_grn}
\title{Gated Residual Network block}
\usage{
layer_grn(
  object,
  hidden_units,
  output_size,
  dropout_rate = NULL,
  use_additional_context = FALSE,
  return_gate = FALSE,
  name = NULL,
  ...
)
}
\description{
Gated Residual Network block
}
