% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer-vsn.R
\name{layer_vsn}
\alias{layer_vsn}
\title{Variable Selection Network block}
\usage{
layer_vsn(
  object,
  input_dim,
  num_inputs,
  hidden_dim,
  dropout_rate = NULL,
  use_context = FALSE,
  name = NULL,
  ...
)
}
\description{
Variable Selection Network block
}
