% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss-negative-log-likelihood.R
\name{loss_negative_log_likelihood}
\alias{loss_negative_log_likelihood}
\title{General negative log likelihood loss function}
\usage{
loss_negative_log_likelihood(...)
}
\arguments{
\item{distribution}{A probability distribution function from \code{tfprobability} package.}
}
\description{
General negative log likelihood loss function
}
\examples{
y_pred <- array(runif(60), c(2, 10, 2))
y_true <- array(runif(20), c(2, 10, 1))

loss_negative_log_likelihood(reduction = 'auto')(y_true, y_pred)
loss_negative_log_likelihood(reduction = 'sum')(y_true, y_pred)
}
