% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-arrays.R
\name{make_arrays}
\alias{make_arrays}
\title{Prepare input/taget arrays for time series models}
\usage{
make_arrays(
  data,
  key,
  index,
  lookback,
  horizon,
  stride = 1,
  shuffle = TRUE,
  sample_frac = 1,
  target,
  numeric = NULL,
  categorical = NULL,
  static = NULL,
  past = NULL,
  future = NULL,
  ...
)
}
\arguments{
\item{lookback}{The length of the context from the past}

\item{horizon}{The forecast length}

\item{stride}{Stride of the moving window}

\item{shuffle}{Shuffle samples. Set \code{FALSE} for the test dataset.}

\item{target}{Target variable(s)}

\item{numeric}{Numeric variables}

\item{categorical}{Categorical variables}

\item{static}{Static variables}
}
\value{
A list of arrays. The maximal possible content embraces eight arrays:
\itemize{
\item \strong{y_past}
\item \strong{X_past_cat}
\item \strong{X_past_num}
\item \strong{y_fut}
\item \strong{X_fut_cat}
\item \strong{X_fut_num}
\item \strong{X_static_cat}
\item \strong{X_static_num}
}
}
\description{
Prepare input/taget arrays for time series models
}
\examples{
library(m5)
library(recipes, warn.conflicts=FALSE)
library(zeallot)
library(dplyr, warn.conflicts=FALSE)
library(data.table, warn.conflicts=FALSE)

# ==========================================================================
#                          PREPARING THE DATA
# ==========================================================================

train <- tiny_m5[date < '2016-01-01']
test  <- tiny_m5[date >= '2016-01-01']

m5_recipe <-
   recipe(value ~ ., data=train) \%>\%
   step_mutate(item_id_idx=item_id, store_id_idx=store_id) \%>\%
   step_integer(item_id_idx, store_id_idx,
                wday, month,
                event_name_1, event_type_1,
                event_name_2, event_type_2,
                zero_based=TRUE) \%>\%
   step_naomit(all_predictors()) \%>\%
   prep()

train <- bake(m5_recipe, train)
test  <- bake(m5_recipe, test)

TARGET      <- 'value'
STATIC      <- c('item_id_idx', 'store_id_idx')
CATEGORICAL <- c('event_name_1', 'event_type_1', STATIC)
NUMERIC     <- c('sell_price', 'sell_price')
KEY         <- c('item_id', 'store_id')
INDEX       <- 'date'
LOOKBACK    <- 28
HORIZON     <- 14
STRIDE      <- LOOKBACK

setDT(train)
setDT(test)

# ==========================================================================
#                           CREATING ARRAYS
# ==========================================================================

train_arrays <-
   make_arrays(
       data        = train,
       key         = KEY,
       index       = INDEX,
       lookback    = LOOKBACK,
       horizon     = HORIZON,
       stride      = STRIDE,
       target      = TARGET,
       static      = STATIC,
       categorical = CATEGORICAL,
       numeric     = NUMERIC
   )

print(names(train_arrays))
print(dim(train_arrays$X_past_num))

test_arrays <-
   make_arrays(
       data        = train,
       key         = KEY,
       index       = INDEX,
       lookback    = LOOKBACK,
       horizon     = HORIZON,
       stride      = STRIDE,
       target      = TARGET,
       static      = STATIC,
       categorical = CATEGORICAL,
       numeric     = NUMERIC
   )

print(names(test_arrays))
print(dim(test_arrays$X_past_num))
}
