% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-arrays.R
\name{make_arrays}
\alias{make_arrays}
\title{Prepare arrays (tensors), which can be used a}
\usage{
make_arrays(
  data,
  key,
  index,
  lookback,
  horizon,
  stride = 1,
  shuffle = TRUE,
  sample_frac = 1,
  target,
  numeric = NULL,
  categorical = NULL,
  static = NULL,
  past = NULL,
  future = NULL,
  keep_all_ids = TRUE,
  ...
)
}
\arguments{
\item{data}{A data.frame object}

\item{key}{Columns, which creates unique time series IDs}

\item{index}{A time-related column}

\item{lookback}{The length of the context from the past}

\item{horizon}{The forecast length}

\item{stride}{Stride of the moving window}

\item{shuffle}{Shuffle samples. Set \code{FALSE} for the test dataset.}

\item{target}{Target variable(s)}

\item{numeric}{Numeric variables}

\item{categorical}{Categorical variables}

\item{static}{Static variables}
}
\description{
Prepare arrays (tensors), which can be used a
}
\examples{
library(m5)
library(recipes)
library(zeallot)

train <- tiny_m5[date < '2016-01-01']
test  <- tiny_m5[date >= '2016-01-01']

m5_recipe <-
   recipe(value ~ ., data=train) \%>\%
   step_mutate(item_id_idx=item_id, store_id_idx=store_id) \%>\%
   step_integer(item_id_idx, store_id_idx,
                wday, month,
                event_name_1, event_type_1,
                event_name_2, event_type_2,
                zero_based=TRUE) \%>\%
   prep()

train <- bake(m5_recipe, train)
test  <- bake(m5_recipe, test)

STATIC      <- c('item_id_idx', 'store_id_idx')
CATEGORICAL <- c('item_id_idx', 'store_id_idx')
NUMERIC     <- c('')

setDT(train)

c(data, index_table) \%<-\%
   make_arrays(
       data = train,
       key = c('item_id', 'store_id'),
       index = 'date',
       lookback = 28,
       horizon = 28,
       static=STATIC,
       categorical=CATEGORICAL
   )

}
